pixiecore: provide a default value for listen-addr flag
As it was, passing dhcp-no-bind with no explicit listen-addr
would fail with an obscure:

listen ip4:17: lookup <nil>: invalid domain name

This used to work ok for me, so I traced back the breaking change
to 3aa769557c8c8a5e606543f0d66b868f1314eef1, which was about
setting Content-Length when serving the kernel/initrd

There, the handling of the listen-addr flag went from a IP address
object with a default value of net.IPv4zero to a plain string with no
default value.

I've reverted that change, which also has the benefit of immediately
catching errors whenever the string passed as the listen-addr is not a
valid IPv4 address, like 192.168.512.10.

OTOH, with the current implementation, one can use domain names
instead of plain IPv4 addresses, which I guess was the reason you
changed it in the first place.  If that's the case, I then suggest
just providing a default value of 0.0.0.0 to listen-addr:

pixiecore/cli/cli.go
-cmd.Flags().StringP("listen-addr", "l", "", "IPv4 address to listen on")
+cmd.Flags().StringP("listen-addr", "l", "0.0.0.0", "IPv4 address to listen on")

Note that I've checked that it also fixes the problem.

[Feel free to edit out the cruft here; Cheers!]