Add partial support for arm64
This is enough support that `pixiecore quick debian testing --arch arm64`
is able to successfully boot a raspberry pi 4 that's using the
experimental edk2 raspberry pi UEFI firmware.

It does not seem to work with the built-in eeprom pxe booting the
raspberry pi 4 includes, but the eeprom pxe implementation is severely
lacking, so I'm not surprised.